ScanScreen styles and subcomps are in new files
