+++ src/navigation/stacks/MapStack.js
@@ -1,7 +1,6 @@
 import React from "react";
 import { createStackNavigator, TransitionPresets } from '@react-navigation/stack';
 
-
 import MapHelpScreen from "../../screens/MapHelpScreen";
 import MapSubStack from './MapSubStack';
